Move css in separate files and a especific folder
