--- src/app.css
@@ -1,374 +1,0 @@
-/* stylelint-disable no-descending-specificity */
-.app {
-  background-color: #4369b2;
-}
-
-body {
-  background-color: #4369b2;
-  overflow-x: hidden;
-  margin-left: -10px;
-  margin-right: -13px;
-}
-
-html {
-  width: 129%;
-  margin-left: -0.5rem;
-}
-
-.countries-list {
-  list-style: none;
-  display: flex;
-  flex-wrap: wrap;
-  margin-top: 0;
-  margin-left: -22px;
-  margin-right: -5rem;
-}
-
-.countries-list li {
-  flex: 11rem;
-  padding-bottom: 1rem;
-  background-color: #4268af;
-}
-
-.countries-list li:nth-child(4n),
-.countries-list li:nth-child(4n+1) {
-  background-color: #5889e6;
-}
-
-.header-wrapper {
-  display: flex;
-  gap: 10rem;
-  background-color: #5889e5;
-  width: 115%;
-  margin-right: 3rem;
-  padding-top: 2rem;
-  padding-bottom: 2rem;
-}
-
-.world-map {
-  width: 11rem;
-  height: 11rem;
-  margin-left: 2rem;
-  opacity: 0.7;
-}
-
-.world-info {
-  display: flex;
-  flex-direction: column;
-  padding-top: 2rem;
-  text-align: center;
-  color: white;
-  transform: scale(1.5);
-}
-
-.world-cases {
-  font-weight: bolder;
-  margin-right: 1rem;
-  font-family: 'lato', sans-serif;
-}
-
-.world-cases-value {
-  font-family: 'lato', sans-serif;
-  font-weight: bolder;
-  margin-right: 1rem;
-  color: rgb(0, 0, 0);
-}
-
-.icon-virus {
-  animation: rotate 1.5s linear infinite;
-  width: 3rem;
-  height: 3rem;
-  margin-left: 2.5rem;
-  margin-top: 1rem;
-  overflow: hidden;
-  transition-duration: 0.5s;
-}
-
-@keyframes rotate {
-  to { transform: rotate(90deg); }
-}
-
-.spinner {
-  display: inline-block;
-  border-radius: 50%;
-  box-shadow: inset -2px 0 0 2px #0bf;
-}
-
-.country-data {
-  display: flex;
-  flex-direction: column;
-  padding-left: -5rem;
-  width: 15rem;
-  gap: 0.5rem;
-}
-
-.country {
-  border: none;
-  background-color: transparent;
-  width: 200px;
-  height: 17rem;
-}
-
-.country-map {
-  opacity: 0.6;
-  margin-right: 1.5rem;
-  transform: scale(1.1);
-  margin-top: 0.5rem;
-}
-
-.country-name {
-  font-weight: bolder;
-  font-size: 1.1rem;
-  transform: scale(1.2);
-  width: 7rem;
-  text-align: center;
-  font-family: 'lato', sans-serif;
-  margin-top: -2rem;
-  color: #1a2a46;
-  margin-left: 4rem;
-}
-
-.total-confirmed {
-  font-weight: bold;
-  color: white;
-  font-family: 'lato', sans-serif;
-  font-size: 1.2rem;
-}
-
-.info-icon {
-  transform: translate(9rem, -8rem);
-  width: 2rem;
-  height: 2rem;
-  margin-left: 5rem;
-}
-
-.table-infos {
-  list-style: none;
-  display: flex;
-  flex-direction: column;
-  gap: 2.5rem;
-  margin-left: 2rem;
-}
-
-.date {
-  float: left;
-  margin-left: 1rem;
-  font-weight: bolder;
-  color: white;
-  font-family: 'lato', sans-serif;
-}
-
-.date-value {
-  float: right;
-  margin-right: 1rem;
-  font-weight: bold;
-  color: white;
-  background-color: transparent !important;
-  font-family: 'lato', sans-serif;
-}
-
-.table-infos li {
-  margin-left: -4rem;
-  height: 5rem;
-  padding: 1rem;
-  color: white;
-  display: table;
-  width: 34rem;
-}
-
-.data-of {
-  color: white;
-  font-weight: 500;
-  font-size: 1.5rem;
-  word-spacing: 10px;
-  margin-left: 1rem;
-  letter-spacing: 1px;
-  font-family: 'lato', sans-serif;
-}
-
-.country-details {
-  display: flex;
-  gap: 1rem;
-}
-
-.data-of-wrapper {
-  margin-bottom: 1rem;
-  margin-top: 1rem;
-  margin-left: 1rem;
-  font-family: 'lato', sans-serif;
-}
-
-.country-total {
-  font-family: 'lato', sans-serif;
-  color: black;
-}
-
-.country-header-info {
-  margin-top: 4rem;
-  display: flex;
-  flex-direction: column;
-  color: white;
-  font-weight: bold;
-  z-index: 11;
-  transform: scale(1.5);
-}
-
-.country-map-header {
-  width: 10rem;
-  height: 10rem;
-  margin-left: 7rem;
-  margin-top: 1.5rem;
-  margin-bottom: 1rem;
-  opacity: 0.7;
-  transform: scale(1.7);
-  z-index: 0;
-}
-
-.table-infos li:nth-child(odd) {
-  background-color: #2d4573;
-}
-
-.table-infos li :nth-child(even) {
-  background-color: #4268af;
-}
-
-.info-ic {
-  text-align: right;
-}
-
-.icon-chev {
-  padding: 0;
-  float: right;
-  width: 2rem;
-  height: 2rem;
-}
-
-.metric {
-  font-weight: bolder;
-  background-color: transparent !important;
-  font-family: 'lato', sans-serif;
-  margin-left: 1rem;
-  width: 2rem;
-  float: left;
-}
-
-.virus {
-  transition: 0.5s;
-  height: 4rem;
-  width: 4rem;
-  animation: cowmove 4s infinite;
-  color: white;
-}
-
-@keyframes cowmove {
-  0% {
-    transform: translateX(10px) rotateY(0deg);
-  }
-
-  49% {
-    transform: translateX(500px) rotateY(0deg);
-  }
-
-  50% {
-    transform: translateX(500px) rotateY(360deg);
-  }
-
-  100% {
-    transform: translateX(100px) rotateY(360deg);
-  }
-}
-
-.metric-value {
-  text-align: center;
-  margin-left: 5rem;
-  font-weight: bold;
-  background-color: transparent !important;
-  font-family: 'lato', sans-serif;
-  padding-top: 0.5rem;
-}
-
-nav {
-  background-color: #4369b2;
-  display: flex;
-  gap: 10rem;
-  margin-bottom: 1rem;
-  max-width: 5rem;
-  padding-top: 1rem;
-}
-
-.location {
-  color: white;
-  font-family: 'lato', sans-serif;
-  font-weight: bold;
-  font-size: 2rem;
-  z-index: 10;
-}
-
-.location-path {
-  display: flex;
-  margin-left: 4rem;
-}
-
-.back {
-  background-color: transparent;
-  color: white;
-  border: none;
-  font-weight: bold;
-  font-size: 2rem;
-  font-family: 'lato', sans-serif;
-}
-
-.icons {
-  display: flex;
-  flex-direction: row;
-  gap: 2rem;
-  margin-left: -4rem;
-}
-
-.search-by {
-  font-weight: bold;
-  margin-right: 2rem;
-  margin-left: 1rem;
-  color: white;
-}
-
-.input-search {
-  border: none;
-  width: 15rem;
-  margin-left: 2rem;
-  height: 1.5rem;
-  background-color: #35548b;
-}
-
-input[type="text"] {
-  font-weight: bold;
-  font-size: 1rem;
-  color: white;
-  padding-right: 1rem;
-  margin-right: 1rem;
-  font-family: 'lato', sans-serif;
-  letter-spacing: 5px;
-  word-spacing: 5px;
-}
-
-::placeholder {
-  font-weight: lighter;
-  font-size: 1rem;
-  color: white;
-  font-family: 'lato', sans-serif;
-}
-
-:focus {
-  outline: 0;
-  outline-color: transparent;
-  outline-style: none;
-}
-
-.form {
-  padding: 1.5rem;
-  width: 102%;
-  margin-top: 0.8rem;
-  background-color: #35548b;
-  margin-bottom: 1rem;
-  margin-left: 1.3rem;
-}
