+++ src/component/CountryInfo.js
@@ -4,6 +4,7 @@ import { BiChevronRightCircle } from 'react-icons/bi';
 import PropTypes from 'prop-types';
 import { useDispatch, useSelector } from 'react-redux';
 import { dataLoading } from '../redux/covid/covid';
+import '../styles/countryInfo.css';
 
 const CountryInfo = ({ current, image }) => {
   const dispatch = useDispatch();
