+++ src/component/Query.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import Proptypes from 'prop-types';
+import '../styles/query.css';
 
 const Query = ({ handleChange }) => (
   <form className="form">
