+++ src/index.js
@@ -2,12 +2,12 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { HashRouter } from 'react-router-dom';
-import App from './App';
+import App from './component/App';
 import store from './redux/configureStore';
 import { loadDataThunk } from './redux/covid/covid';
+import './styles/index.css';
 
 store.dispatch(loadDataThunk());
-// store.dispatch(loadHistoryThunk('france'));
 
 ReactDOM.render(
   <React.StrictMode>
