+++ src/styles/app.css
@@ -1,0 +1,4 @@
+
+.app {
+  background-color: #4369b2;
+}
