+++ src/styles/country.css
@@ -1,0 +1,46 @@
+.country-data {
+  display: flex;
+  flex-direction: column;
+  padding-left: -5rem;
+  width: 15rem;
+  gap: 0.5rem;
+}
+
+.country {
+  border: none;
+  background-color: transparent;
+  width: 200px;
+  height: 17rem;
+}
+
+.country-map {
+  opacity: 0.6;
+  margin-right: 1.5rem;
+  transform: scale(1.1);
+  margin-top: 0.5rem;
+}
+
+.country-name {
+  font-weight: bolder;
+  font-size: 1.1rem;
+  transform: scale(1.2);
+  width: 7rem;
+  text-align: center;
+  font-family: 'lato', sans-serif;
+  margin-top: -2rem;
+  color: #1a2a46;
+  margin-left: 4rem;
+}
+
+.total-confirmed {
+  font-weight: bold;
+  color: white;
+  font-family: 'lato', sans-serif;
+  font-size: 1.2rem;
+}
+
+.info-icon {
+  transform: translate(9rem, -8rem);
+  width: 2rem;
+  height: 2rem;
+}
