+++ src/styles/countryInfo.css
@@ -1,0 +1,144 @@
+.table-infos {
+  list-style: none;
+  display: flex;
+  flex-direction: column;
+  gap: 2.5rem;
+  margin-left: 2rem;
+}
+
+.date {
+  float: left;
+  margin-left: 1rem;
+  font-weight: bolder;
+  color: white;
+  font-family: 'lato', sans-serif;
+}
+
+.date-value {
+  float: right;
+  margin-right: 1rem;
+  font-weight: bold;
+  color: white;
+  background-color: transparent !important;
+  font-family: 'lato', sans-serif;
+}
+
+.table-infos li {
+  margin-left: -4rem;
+  height: 5rem;
+  padding: 1rem;
+  color: white;
+  display: table;
+  width: 34rem;
+}
+
+.data-of {
+  color: white;
+  font-weight: 500;
+  font-size: 1.5rem;
+  word-spacing: 10px;
+  margin-left: 1rem;
+  letter-spacing: 1px;
+  font-family: 'lato', sans-serif;
+}
+
+.country-details {
+  display: flex;
+  gap: 1rem;
+}
+
+.data-of-wrapper {
+  margin-bottom: 1rem;
+  margin-top: 1rem;
+  margin-left: 1rem;
+  font-family: 'lato', sans-serif;
+}
+
+.country-total {
+  font-family: 'lato', sans-serif;
+  color: black;
+}
+
+.country-header-info {
+  margin-top: 4rem;
+  display: flex;
+  flex-direction: column;
+  color: white;
+  font-weight: bold;
+  z-index: 11;
+  transform: scale(1.5);
+}
+
+.country-map-header {
+  width: 10rem;
+  height: 10rem;
+  margin-left: 7rem;
+  margin-top: 1.5rem;
+  margin-bottom: 1rem;
+  opacity: 0.7;
+  transform: scale(1.7);
+  z-index: 0;
+}
+
+.table-infos li:nth-child(odd) {
+  background-color: #2d4573;
+}
+
+.table-infos li :nth-child(even) {
+  background-color: #4268af;
+}
+
+.info-ic {
+  text-align: right;
+}
+
+.icon-chev {
+  padding: 0;
+  float: right;
+  width: 2rem;
+  height: 2rem;
+}
+
+.metric {
+  font-weight: bolder;
+  background-color: transparent !important;
+  font-family: 'lato', sans-serif;
+  margin-left: 1rem;
+  width: 2rem;
+  float: left;
+}
+
+.virus {
+  transition: 0.5s;
+  height: 4rem;
+  width: 4rem;
+  animation: cowmove 4s infinite;
+  color: white;
+}
+
+@keyframes cowmove {
+  0% {
+    transform: translateX(10px) rotateY(0deg);
+  }
+
+  49% {
+    transform: translateX(500px) rotateY(0deg);
+  }
+
+  50% {
+    transform: translateX(500px) rotateY(360deg);
+  }
+
+  100% {
+    transform: translateX(100px) rotateY(360deg);
+  }
+}
+
+.metric-value {
+  text-align: center;
+  margin-left: 5rem;
+  font-weight: bold;
+  background-color: transparent !important;
+  font-family: 'lato', sans-serif;
+  padding-top: 0.5rem;
+}
