+++ src/styles/index.css
@@ -1,0 +1,15 @@
+/* *{
+    outline: 1px solid red;
+} */
+
+body {
+  background-color: #4369b2;
+  overflow-x: hidden;
+  margin-left: -10px;
+  margin-right: -13px;
+}
+
+html {
+  width: 129%;
+  margin-left: -0.5rem;
+}
